requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.21.4
idna==3.4
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tweepy==4.10.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1
